+++ src/renderer/skills/Hub.ts
@@ -1,10 +1,17 @@
 import {EventEmitter} from "events";
 import Skill from './Skill';
-import TTSController from '../TTSController';
-import AsyncToken from '../AsyncToken';
-import BingTTSController from '../microsoft/BingTTSController';
+import HotwordController, { HotwordResult } from '../HotwordController';
+import SnowboyController from '../snowboy/SnowboyController';
+import STTController from '../STTController';
+import BingSpeechApiController from '../microsoft/BingSpeechApiController';
 import NLUController, { NLUIntentAndEntities } from '../NLUController';
 import LUISController from '../microsoft/LUISController';
+import TTSController from '../TTSController';
+import BingTTSController from '../microsoft/BingTTSController';
+import AsyncToken from '../AsyncToken';
+import PixijsManager from '../pixijs/PixijsManager';
+import RomManager from '../rom/RomManager';
+import WwMusicController from '../ww/WwMusicController';
 
 import ClockSkill from './ClockSkill';
 import JokeSkill from './JokeSkill';
@@ -28,6 +35,7 @@ export default class Hub extends EventEmitter {
 
     constructor(options?: HubOptions) {
         super ();
+        console.log(`HUB: CONSTRUCTOR!!`);
         this.skillMap = new Map<string, Skill>();
         this.launchIntentMap = new Map<string, Skill>();
         this.audioContext = new AudioContext();
@@ -43,6 +51,10 @@ export default class Hub extends EventEmitter {
     static Instance(options?: HubOptions)
     {
         return this._instance || (this._instance = new this(options));
+    }
+
+    init(): void {
+
     }
 
     tick(): void {
@@ -96,5 +108,87 @@ export default class Hub extends EventEmitter {
             .catch((error: any) => {
                 console.log(error);
             });
+    }
+
+    startNLU(utterance: string) {
+        const nluController: NLUController = new LUISController();
+    
+        let t: AsyncToken<NLUIntentAndEntities> = nluController.getIntentAndEntities(utterance);
+    
+        t.complete
+            .then((intentAndEntities: NLUIntentAndEntities) => {
+                console.log(`NLUIntentAndEntities: `, intentAndEntities);
+                RomManager.Instance().onNLU(intentAndEntities, utterance);
+                Hub.Instance().handleLaunchIntent(intentAndEntities, utterance);
+            })
+            .catch((error: any) => {
+                console.log(error);
+            });
+    }
+
+    startRecognizer() {
+        PixijsManager.Instance().eyeShowHighlight();
+        setTimeout(() => {
+            this.startHotword();
+        }, 1000);
+        // const speechController: STTController = new BingSpeechApiController();
+        // let t: AsyncToken<string> = speechController.RecognizerStart({recordDuration: 3000});
+    
+        // t.on('Listening', () => {
+        //     //console.log(`renderer: startRecognizer: on Listening`);
+        // });
+    
+        // t.on('RecognitionEndedEvent', () => {
+        //     //console.log(`renderer: startRecognizer: on RecognitionEndedEvent`);
+        // });
+    
+        // t.on('Recording_Stopped', () => {
+        //     //console.log(`renderer: startRecognizer: on Recording_Stopped`);
+        //     this.startHotword();
+        // });
+    
+        // t.complete
+        //     .then((utterance: string) => {
+        //         console.log(`Utterance: ${utterance}`);
+        //         if (utterance) {
+        //             RomManager.Instance().onUtterance(utterance);
+        //             this.startNLU(utterance);
+        //         }
+        //     })
+        //     .catch((error: any) => {
+        //         console.log(error);
+        //     });
+    
+    }
+
+    startHotword() {
+        console.log(`START HOTWORD`);
+        console.trace();
+        const hotwordController: HotwordController = new SnowboyController();
+        let t: AsyncToken<HotwordResult> = hotwordController.RecognizerStart({sampleRate: 16000});
+        PixijsManager.Instance().eyeBlink();
+        PixijsManager.Instance().eyeShowHighlight(false);
+    
+        t.on('Listening', () => {
+            //console.log(`renderer: startHotword: on Listening`);
+        });
+    
+        t.on('hotword', () => {
+            //console.log(`renderer: startHotword: on hotword: `, eyeInstance);
+            RomManager.Instance().onHotword();
+        });
+    
+        t.complete
+            .then((result: HotwordResult) => {
+                // console.log(`HotWord: result:`, result);
+                this.startRecognizer();
+            })
+            .catch((error: any) => {
+                console.log(error);
+            });
+    }
+
+    startMusic() {
+        const musicController = new WwMusicController();
     }
 }
